+++ mobile/android/app/src/main/java/kitchen/cegin/app/MainApplication.kt
@@ -22,6 +22,7 @@ class MainApplication : Application(), ReactApplication {
       packageList =
         PackageList(this).packages.apply {
           add(MaterialYouPackage())
+          add(ExactAlarmPackage())
           // Packages that cannot be autolinked yet can be added manually here, for example:
           // add(MyReactNativePackage())
         }